web/requirements.txt
flask
werkzeug
apscheduler
gunicorn
python-dotenv
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask_sqlalchemy